CreateVertical

+++ src/modules/managers/pair_manager.h
@@ -19,8 +19,8 @@ public:
 
   bool GetUsePairData();
   void SetPairData(QString character, QString emote, int selfOffset, int pairOffset, bool flipped, int scale, int vertical);
-  double GetOffsetSelf();
-  double GetOffsetOther();
+  int GetOffsetSelf();
+  int GetOffsetOther();
   double GetVerticalOffset() {return (double)mVerticalOffset; }
   double GetScaleoffset() {return (double)mScaleOffset / (double)1000.0f; }
